Take into account the views of the substates
In the experimentation I made, I realised the `$viewContentLoaded` event is fired once per declared state in the state hierarchy, plus one extra time because of the `""` state, which is the root to all states. Thus, we need to figure out the amount of times the event will be fired and count them until all are accounted for.